% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ClimaRep-package.R
\docType{package}
\name{ClimaRep-package}
\alias{ClimaRep-package}
\alias{ClimaRep}
\title{ClimaRep: ClimaRep: Estimating Climate Representativeness}
\description{
The **ClimaRep** package offers tools to estimate the climate representativeness of defined areas and quantifies and analyzes its transformation under future climate change scenarios.
}
\author{
\strong{Maintainer}: Mario Mingarro Lopez \email{mario_mingarro@mncn.csic.es} (\href{https://orcid.org/0000-0003-3977-7944}{ORCID})

}
\keyword{internal}
