% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mh_rep_ch.R
\name{mh_rep_ch}
\alias{mh_rep_ch}
\title{Multivariate Temporal Climate Representativeness Change Analysis}
\usage{
mh_rep_ch(
  polygon,
  col_name,
  present_climate_variables,
  future_climate_variables,
  study_area,
  th = 0.95,
  model,
  year,
  dir_output = file.path(tempdir(), "ClimaRep"),
  save_raw = FALSE
)
}
\arguments{
\item{polygon}{An \code{sf} object containing the definied areas.}

\item{col_name}{\code{Character}. Name of the column in the \code{polygon} object that contains unique identifiers for each polygon.}

\item{present_climate_variables}{\code{SpatRaster}. A raster stack of climate variables representing \strong{present} conditions (preferably not strongly correlated).}

\item{future_climate_variables}{\code{SpatRaster}. A raster stack containing the \emph{same} climate variables as \code{present_climate_variables} but representing \strong{future} projected conditions. Must have the \strong{same extent} and resolution as \code{present_climate_variables}.}

\item{study_area}{\code{sf}. A single polygon defining the overall boundary of the climate space.}

\item{th}{\code{Numeric} (0-1).Percentile threshold used to define representativeness. Cells with a Mahalanobis distance below or equal to the \code{th} are classified as representative (default: 0.95).}

\item{model}{\code{Character.} Name or identifier of the climate model used for future projections (e.g., "MIROC6"). This parameter is mandatory and used in output filenames and subdirectory names.}

\item{year}{\code{Character.} Projection year or period for future climate data (e.g., "2070"). This parameter is mandatory and used in output filenames and subdirectory names.}

\item{dir_output}{\code{Character}. Path to the directory where output files will be saved. The function will create subdirectories within this path.}

\item{save_raw}{\code{Logical.} If \code{TRUE}, saves the intermediate continuous Mahalanobis distance rasters calculated for each polygon before binary classification. The final binary classification rasters are always saved (default: FALSE).}
}
\value{
Invisibly returns NULL. Writes the following outputs to disk within subdirectories of \code{dir_output}:
\itemize{
\item Classification (\code{.tif} ) change raster: Change category rasters (\code{0} for \strong{Non-representative}, \code{1} for \strong{Retained}, \code{2} for \strong{Lost} and \code{3} for \strong{Novel}) for each input \code{polygon} are saved in the \verb{Change/} subdirectory.
\item Visualization (\code{.jpeg}) maps: Image files visualizing the change classification results for each \code{polygon} are saved in the \verb{Charts/} subdirectory.
\item Raw Mahalanobis Distance Rasters: \emph{Optionally} saved as \code{.tif} files in the \verb{Mh_Raw_Pre/} and \verb{Mh_Raw_Fut/} subdirectory if \code{save_raw = TRUE}.
}
}
\description{
Calculates Mahalanobis-based climate representativeness for input polygons across two time periods (present and future) within a defined study area. The function identifies areas of climate representativeness persistence, loss, or gain.
Representativeness is assessed by comparing the multivariate climate conditions of each cell to the reference climate space defined by the climate conditions \emph{within} each specific input polygon in the \emph{present} period, scaled by the overall climate variability across both time periods within the study area.
}
\details{
This function extends the multivariate analysis approach used in \code{mh_rep} to assess changes in climate representativeness (or \strong{forward climate analogs}) over time.
While \code{mh_rep()} calculates representativeness relative to the mean and covariance of climate conditions \emph{without change}, \code{mh_rep_ch()} adapts this by using the mean from the present polygon but a covariance matrix derived from the overall climate space across both present and future periods combined.

The process involves comparing representativeness between \strong{present} and \strong{future} periods within a defined climate scpace (\code{study_area}) for areas defined by input polygons (\code{polygon}).
Key workflow steps include:
\enumerate{
\item Ensure all spatial inputs (\code{polygon}, \code{present_climate_variables}, \code{future_climate_variables}, \code{study_area}) share the same Coordinate Reference System (CRS), using the CRS of \code{present_climate_variables} as the reference.
\item Calculate the multivariate covariance matrix using climate data from \emph{all cells within the study area} for \emph{both} present and future time periods combined. This captures the overall climate variability against which distances are scaled.
\item For each polygon in the \code{polygon} object:
\itemize{
\item Calculate the multivariate mean climate using climate data \emph{only from within that polygon} for the \strong{present} time period. This defines the center of the reference climate space for the polygon.
\item Calculate the Mahalanobis distance for all cells within the study area, relative to the polygon's \strong{present} conditions (calculated in the previous step) and the overall \strong{present} and \strong{future} covariance matrix (calculated in step 2).
This results in a Mahalanobis distance raster for the present period and a Mahalanobis distance raster for the future period.
\item Determine a threshold (\code{th}) based on the \code{th} percentile of the Mahalanobis distances calculated \emph{only from within the present polygon}.
\item Classify, based on the threshold (\code{th}), all cells within the \code{study_area} for both present and future periods as Representative or forward climate analogs (mh < \code{th}) or "Non-Representative" (mh > \code{th}).
}
\item Compares the binary representativeness status of each cell between the present and future periods and determines cells where conditions:
\itemize{
\item \code{0} \strong{Not Represented}
\item \code{0} \strong{Retained} Are present in both currently and future.
\item \code{0} \strong{Lost} Are present currently, but not present in the future.
\item \code{0} \strong{Novel} Are present in the future, but not present currently.
\item Output Generation: Saves the classification raster (\code{tif}) for each polygon and generates a corresponding visualization map (\code{jpeg}). These are saved within the specified output directory (\code{dir_output}).
All files are saved using the \code{model} and \code{year} parameters in filenames for identification.
}}

It is important to note that Mahalanobis distance assumes multivariate normality and is sensitive to collinearity among variables.
While the covariance matrix accounts for correlations, it is strongly recommended that the climate variables (\code{present_climate_variables} and \code{future_climate_variables}) are not strongly correlated.
Consider performing a collinearity analysis beforehand, perhaps using the \code{vif_filter} function from this package.
}
\examples{
\dontrun{
library(terra)
library(sf)
set.seed(2458)
n_cells <- 100 * 100
r_clim_present <- terra::rast(ncols = 100, nrows = 100, nlyrs = 7)
values(r_clim_present) <- c(
  (rowFromCell(r_clim_present, 1:n_cells) * 0.2 + rnorm(n_cells, 0, 3)),
  (rowFromCell(r_clim_present, 1:n_cells) * 0.9 + rnorm(n_cells, 0, 0.2)),
  (colFromCell(r_clim_present, 1:n_cells) * 0.15 + rnorm(n_cells, 0, 2.5)),
  (colFromCell(r_clim_present, 1:n_cells) +
    (rowFromCell(r_clim_present, 1:n_cells)) * 0.1 + rnorm(n_cells, 0, 4)),
  (colFromCell(r_clim_present, 1:n_cells) /
    (rowFromCell(r_clim_present, 1:n_cells)) * 0.1 + rnorm(n_cells, 0, 4)),
  (colFromCell(r_clim_present, 1:n_cells) *
    (rowFromCell(r_clim_present, 1:n_cells) + 0.1 + rnorm(n_cells, 0, 4)),
  (colFromCell(r_clim_present, 1:n_cells) *
    (colFromCell(r_clim_present, 1:n_cells) + 0.1 + rnorm(n_cells, 0, 4)))
names(r_clim_present) <- c("varA", "varB", "varC", "varD", "varE", "varF", "varG")
terra::crs(r_clim_present) <- "EPSG:4326"
terra::plot(r_clim_present)
r_clim_present_filtered <- vif_filter(r_clim_present, th = 5)
hex_grid <- sf::st_sf(
  sf::st_make_grid(
    sf::st_as_sf(
      terra::as.polygons(
        terra::ext(r_clim_present)
      ),
    square = FALSE
  )
)
sf::st_crs(hex_grid) <- "EPSG:4326"
polygons <- hex_grid[sample(nrow(hex_grid), 2), ]
polygons$name <- c("Pol_1", "Pol_2")
sf::st_crs(polygons) <- sf::st_crs(hex_grid)
study_area_polygon <- sf::st_as_sf(as.polygons(terra::ext(r_clim_present)))
sf::st_crs(study_area_polygon) <- "EPSG:4326"
terra::plot(r_clim_present[[1]])
terra::plot(polygons, add = TRUE, color = "transparent", lwd = 3)
terra::plot(study_area_polygon, add = TRUE, col = "transparent", lwd = 3, border = "red")
mh_rep(
  polygon = polygons,
  col_name = "name",
  climate_variables = r_clim_present_filtered,
  th = 0.95,
  dir_output = file.path(tempdir(), "ClimaRep"),
  save_raw = TRUE
)
}
}
